Stripe ok

+++ db/migrate/20140512101012_remove_price_from_course.rb
@@ -1,5 +1,5 @@
 class RemovePriceFromCourse < ActiveRecord::Migration
   def change
-    remove_column :courses, :price, :integer
+    remove_column :courses, :price, :decimal
   end
 end
