+++ db/migrate/20140512101044_add_price_from_course.rb
@@ -1,5 +1,5 @@
 class AddPriceFromCourse < ActiveRecord::Migration
   def change
-    add_column :courses, :price, :decimal
+    add_column :courses, :price, :integer
   end
 end
